+++ apps/web/components/utils.tsx
@@ -70,7 +70,6 @@ export async function updateProblem(problemId: string, data: any) {
     });
     return problem;
   } catch (e) {
-    console.log(e);
     return null;
   }
 }
@@ -237,7 +236,6 @@ export async function createTrack(data: {
     }
     return track;
   } catch (e) {
-    console.log(e);
     return new Error("Failed to create track");
   }
 }
@@ -273,19 +271,31 @@ export async function getAllCategories() {
 
 export async function AddDatatoAlgolia({ trackId }: { trackId: string }) {
   const client = algoliasearch(process.env.AlGOLIA_APP_ID!, process.env.AlGOLIA_Admin_API!);
+  const notion = new NotionAPI();
   try {
     const index = client.initIndex("DailyCode");
     const track = await getTrack(trackId);
-    const notion = new NotionAPI();
     const data = await Promise.all(
       track?.problems.map(async (problem: any) => {
         const notionDocId = problem.notionDocId;
         const notionPage = await notion.getPage(notionDocId);
         const titles = Object.values(notionPage.block)
-          .map((block) => block?.value?.properties?.title[0][0])
+          .map((block) => {
+            const title = block?.value?.properties?.title;
+            if (title && title[0] && title[0][0]) {
+              return title[0][0];
+            } else {
+              const source = block?.value?.properties?.source;
+              if (source && source[0] && source[0][0]) {
+                return source[0][0];
+              } else {
+                return null;
+              }
+            }
+          })
           .filter((title) => title);
 
-        problem["titles"] = titles;
+        problem["titles"] = titles.flat();
         problem["trackId"] = track.id;
         problem["trackTitle"] = track.title;
         problem["image"] = track.image;
@@ -306,5 +316,26 @@ export async function AddDatatoAlgolia({ trackId }: { trackId: string }) {
     return data;
   } catch (e) {
     return null;
+  }
+}
+
+export async function getSearch(searchText: string) {
+  const client = algoliasearch(process.env.AlGOLIA_APP_ID!, process.env.AlGOLIA_Admin_API!);
+  try {
+    const index = client.initIndex("DailyCode");
+    const data = await index.search(searchText);
+    let hitsData = data.hits.map((hit: any) => {
+      return {
+        id: hit.id,
+        notionDocId: hit.notionDocId,
+        title: hit.title,
+        trackId: hit.trackId,
+        trackTitle: hit.trackTitle,
+        image: hit.image,
+      };
+    });
+    return hitsData;
+  } catch (e) {
+    return e;
   }
 }
